requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
boto3==1.39.9
botocore==1.39.9
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
fastavro==1.11.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.26.0
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
logfire-api==3.25.0
markdown-it-py==3.0.0
mcp==1.12.0
mdurl==0.1.2
mistralai==1.9.2
multidict==6.6.3
oauthlib==3.3.1
openai==1.97.0
opentelemetry-api==1.35.0
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.4.4
pydantic-ai-slim==0.4.4
pydantic-evals==0.4.4
pydantic-graph==0.4.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.1
tenacity==8.5.0
tokenizers==0.21.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
